pyproject.toml
[build-system]
requires = [
   "setuptools",
   "wheel",
   "build",
   "setuptools_scm"
]
build-backend = "setuptools.build_meta"

[project]
name = "pyeyes"
dynamic = ["version"]

authors = [
  { name="Zachary Shah", email="zshah9@stanford.edu" },
  { name="Yonatan Urman", email="yurman@stanford.edu" },
]
description = "A GUI viewer for high-dimensional image visualization in Python"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "numpy>=1.23",
    "matplotlib",
    "holoviews>=1.20.0",
    "bokeh>=3.6.2",
    "nibabel",
    "scipy",
    "jupyter_bokeh",
    "scikit-image",
]

[project.urls]
Homepage = "https://github.com/zachary-shah/pyeyes"
Repository = "https://github.com/zachary-shah/pyeyes"
"Bug Tracker" = "https://github.com/zachary-shah/pyeyes/issues"
Documentation = "https://zachary-shah.github.io/pyeyes/"

[project.optional-dependencies]
docs = [
    "mkdocs-material",
    "mkdocstrings[python]",
]
dev = [
    "pre-commit",
    "black",
]
test = [
    "pytest>=6.0",
    "pytest-playwright>=0.4.0",
    "pytest-timeout",
]

[tool.setuptools]
include-package-data = true
packages = ["pyeyes"]
package-dir = {"" = "src"}

[tool.setuptools.package-data]
pyeyes = ["cmap/*.csv"]

[tool.setuptools_scm]
version_scheme = "guess-next-dev"
local_scheme = "no-local-version"
